# openApiSchema.yml
openapi: 3.0.0
info:
  title: 股票数据管理API
  description: |
    提供股票数据管理、用户认证和用户数据管理功能的API接口。
    包括股票数据的CRUD操作、用户注册登录、股票选择等功能。
  version: '3.0.0'
  contact:
    name: Stockstock
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: https://api.example.com
    description: 生产环境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 使用JWT令牌进行认证，格式为 "token"
  schemas:
...
    AddStockRequest:
      type: array
      items:
        type: string
        description: 股票代码
      description: 要添加的股票ID数组
      required:
        - stockId

    # 模型预测相关模式
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: 错误消息
        code:
          type: string
          description: 错误代码
    RowsAffectedResponse:
      type: object
      properties:
        rowsAffected:
          type: integer
          description: 受影响的行数
paths:
...
  /persons/stocks:
    get:
      tags:
        - 用户股票管理
      summary: 获取用户的股票列表
      description: 获取当前用户选择的股票列表
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  stocks:
                    type: array
                    items:
                      type: string
                    description: 用户选择的股票代码列表
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - 用户股票管理
      summary: 添加股票到用户列表
      description: 向当前用户的选择列表添加股票
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStockRequest'
      responses:
        '200':
          description: 添加成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rowsAffected:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /persons/stocks/{stockId}:
    delete:
      tags:
        - 用户股票管理
      summary: 从用户列表移除股票
      description: 从当前用户的选择列表中移除特定股票
      security:
        - bearerAuth: []
      parameters:
        - name: stockId
          in: path
          required: true
          schema:
            type: string
          description: 要移除的股票ID
      responses:
        '200':
          description: 移除成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rowsAffected:
                    type: integer
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /stocks:
    get:
      tags:
        - 股票数据
      summary: 获取所有股票信息
      description: 获取数据库中所有股票记录
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - 股票数据
      summary: 创建股票记录
      description: 添加新的股票数据记录(需管理员权限)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockCreate'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  rowsAffected:
                    type: integer
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - 股票数据
      summary: 删除所有股票记录
      description: 删除数据库中所有股票记录(需管理员权限)
      security:
        - bearerAuth: []
      responses:
        '204':
          description: 删除成功
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stocks/byId/{id}:
    get:
      tags:
        - 股票数据
      summary: 按ID查询股票
      description: 获取指定股票代码的所有记录
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 股票代码
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stocks/byDate/{date}:
    get:
      tags:
        - 股票数据
      summary: 按日期查询股票
      description: 获取指定日期的所有股票记录
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: 日期
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Stock'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /stocks/{id}/{date}:
    get:
      tags:
        - 股票数据
      summary: 按ID和日期查询股票
      description: 获取特定股票代码和特定日期的记录
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 股票代码
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: 日期
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stock'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - 股票数据
      summary: 更新股票记录
      description: 更新特定股票代码和特定日期的记录(需管理员权限)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 股票代码
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: 日期
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StockCreate'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsAffectedResponse'
        '400':
          description: 请求参数错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - 股票数据
      summary: 删除股票记录
      description: 删除特定股票代码和特定日期的记录(需管理员权限)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 股票代码
        - name: date
          in: path
          required: true
          schema:
            type: string
          description: 日期
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RowsAffectedResponse'
        '401':
          description: 未认证
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: 权限不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: 资源不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: 服务器内部错误
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # 雪球API股票K线数据
...